scripts/nexusmods-sync/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyGithub>=2.1.0
cloudscraper>=1.2.71
